consolidate file utilts
